refactor : scheduleRepository의 CQS패턴 적용
